Move Authentication to a Custom Hook
Dashboard.tsx only focuses on UI rendering now and a new hook handles the authentication logic (Separation of Concerns). This also makes the code cleaner and improve reusability if we need authentication elsewhere
